mud_backend/urls.py
"""mud_backend URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/2.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from graphene_django.views import GraphQLView
from rest_framework import routers

from room.api import RoomViewSet
from player.api import PlayerViewSet
from map.api import MapViewSet

router = routers.DefaultRouter()


#this will be /api/room
router.register(r'room',RoomViewSet)

#this will correspond to /api/player
router.register(r'player',PlayerViewSet)

#this will correspond to /api/map
router.register(r'map', MapViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('graphql/', GraphQLView.as_view(graphiql=True)),
    path('api/', include(router.urls)),
]
